feat: improve overall feature + UI of Settings page - add delete account dialog - add loading bar while deleting account - improve UI of Settings page

+++ app/src/main/java/com/yeonkims/realnoteapp/logic/viewmodels/auth/AuthViewModel.kt
@@ -1,14 +1,17 @@
 package com.yeonkims.realnoteapp.logic.viewmodels.auth
 
 import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.Transformations
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.yeonkims.realnoteapp.data.enums.AuthState
 import com.yeonkims.realnoteapp.data.repositories.NoteRepository
 import com.yeonkims.realnoteapp.data.repositories.UserRepository
 import com.yeonkims.realnoteapp.logic.viewmodels.AlertViewModel
+import com.yeonkims.realnoteapp.util.dev_tools.Logger
 import kotlinx.coroutines.launch
 import javax.inject.Inject
+import javax.inject.Singleton
 
 class AuthViewModel @Inject constructor(
     private val userRepository: UserRepository,
@@ -17,6 +20,8 @@ class AuthViewModel @Inject constructor(
 ) : ViewModel() {
 
     val currentUser = userRepository.getCurrentUser()
+
+    var isLoading: MutableLiveData<Boolean> = MutableLiveData(false)
 
     var authState : MutableLiveData<AuthState> = MutableLiveData(null)
 
@@ -36,8 +41,13 @@ class AuthViewModel @Inject constructor(
     }
 
     fun deleteAccount() {
+        if(isLoading.value == true) {
+            return
+        }
         viewModelScope.launch {
+            isLoading.value = true
             userRepository.deleteAccount(currentUser.value!!.email).addOnCompleteListener { task ->
+                isLoading.value = false
                 if(task.isSuccessful) {
                     logout()
                     alertViewModel.recordAlertMessage("Your account is now deleted.")
